Merge frontend and main

+++ server/src/main/java/com/ldg/main/Controllers/UserController.java
@@ -3,7 +3,11 @@ package com.ldg.main.Controllers;
 import java.util.List;
 import java.util.Optional;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,7 +19,10 @@ import org.springframework.web.bind.annotation.RestController;
 import com.ldg.main.Services.UserService;
 import com.ldg.main.Models.User;
 import com.ldg.main.Repository.UserRepository;
-
+import com.ldg.main.Policies.UserPolicy;
+import com.ldg.main.Models.UserDetailsImpl; 
+import com.ldg.main.payload.request.ChangePasswordRequest;
+import com.ldg.main.payload.request.ChangeUserRequest;
 @RestController
 @RequestMapping("/api/users")
 public class UserController {
@@ -25,36 +32,51 @@ public class UserController {
     @Autowired
     private UserService userService;
     private Optional<User> user;
+    private Authentication auth;
+
+    @Autowired
+    private UserPolicy policy;
 
     @GetMapping
     public List<User> index() {
-        List<User> Users = userRepository.findAll();
-        return Users;
+        auth = SecurityContextHolder.getContext().getAuthentication();
+        if(policy.index(auth))
+        {
+            List<User> Users = userRepository.findAll();
+            return Users;
+        }
+        return null;
     }
 
     @GetMapping("/{user}")
     public User show(@PathVariable(value = "user") Long ID) {
+        if(policy.show())
+        {
         user = userRepository.findById(ID);
         if (user.isPresent())
             return user.get();
-        else
-            return null;
+        }
+        return null;
     }
 
-    @PutMapping("/{user}")
-    public User update(@PathVariable(value = "user") Long ID) {
-        user = userRepository.findById(ID);
-        return new User();
-    }
-
-    @PostMapping
-    public User store(@RequestBody User user) {
-        return userService.saveUser(user);
+    @PutMapping("/{user}/{role}")
+    public boolean changeRole(@PathVariable(value = "user")Long ID,@PathVariable(value="role")long roleID)
+    {
+        if(policy.isAdmin(auth))
+        {
+            if(userRepository.updateRole(roleID,ID)!=0)
+                return true;
+        }
+        return false;
     }
 
     @DeleteMapping("/{user}")
     public boolean destroy(@PathVariable(value = "user") Long ID) {
-        userRepository.delete(this.show(ID));
-        return true;
+        if(policy.delete(auth,ID))
+        {
+            userRepository.delete(this.show(ID));
+            return true;
+        }
+        return false;
     }
 }
